+++ static_report/src/components/shared/Tables/TableList/TableListItem.tsx
@@ -1,70 +1,187 @@
-import { ReactNode } from 'react';
-import {
-  AccordionButton,
-  AccordionButtonProps,
-  Flex,
-  Icon,
-  Tooltip,
-  Text,
-} from '@chakra-ui/react';
-import { FiAlertCircle, FiGrid } from 'react-icons/fi';
+import { Flex, Grid, Text, GridItem, Icon, Link } from '@chakra-ui/react';
+import { FiChevronDown, FiChevronUp } from 'react-icons/fi';
 
-export interface TableListItemProps extends AccordionButtonProps {
+import {
+  TableAccordionWrapper,
+  TableItemName,
+  TableItemDescription,
+} from './TableListItemDecorations';
+import { CRTableListColumnsSummary } from './CRTableListItem/CRTableListColumnsSummary';
+import { CRTableListDeltaSummary } from './CRTableListItem/CRTableListDeltaSummary';
+
+import type {
+  Comparable,
+  SaferTableSchema,
+  Selectable,
+} from '../../../../types';
+import {
+  formatColumnValueWith,
+  formatNumber,
+} from '../../../../utils/formatters';
+import { SRTableListColumnLabel } from './SRTableListItem/SRTableListColumnLabel';
+import { getIconForColumnType } from '../../Columns/utils';
+import { NoData } from '../../Layouts';
+import { AssertionLabel } from '../../Assertions/AssertionLabel';
+import { CRTableListAssertions } from './CRTableListItem/CRTableListAssertions';
+import { getReportAggregateAssertions } from '../utils';
+import { tableListGridTempCols } from '../../../../utils/layout';
+
+interface Props extends Selectable, Comparable {
   isExpanded: boolean;
-  children: ReactNode;
+  baseTableDatum?: SaferTableSchema;
+  targetTableDatum?: SaferTableSchema;
 }
 
 export function TableListItem({
   isExpanded,
-  children,
-  ...props
-}: TableListItemProps) {
+  baseTableDatum,
+  targetTableDatum,
+  onSelect,
+  singleOnly,
+}: Props) {
+  const fallbackTable = baseTableDatum || targetTableDatum;
+  const tableName = fallbackTable?.name;
+  const description =
+    baseTableDatum?.description || targetTableDatum?.description;
+  //SR vars
+  const columns = Object.keys(baseTableDatum?.columns || {}).map((key) => key);
+
+  const { failed: baseFailed, total: baseTotal } = getReportAggregateAssertions(
+    baseTableDatum?.piperider_assertion_result,
+    baseTableDatum?.dbt_assertion_result,
+  );
+  const { failed: targetFailed, total: targetTotal } =
+    getReportAggregateAssertions(
+      targetTableDatum?.piperider_assertion_result,
+      targetTableDatum?.dbt_assertion_result,
+    );
+
+  if (!fallbackTable) {
+    return <NoData />;
+  }
   return (
-    <AccordionButton bgColor="white" borderRadius="md" {...props}>
-      <Flex
-        direction="column"
-        gap={4}
-        py="10px"
-        maxH={isExpanded ? '135px' : '90px'}
+    <TableAccordionWrapper
+      isExpanded={isExpanded}
+      data-cy="cr-table-overview-btn"
+    >
+      <Grid
+        templateColumns={tableListGridTempCols}
+        justifyItems="flex-start"
+        width="calc(900px - 30px)"
+        position={'relative'}
+        rowGap={3}
       >
-        {children}
-      </Flex>
-    </AccordionButton>
-  );
-}
-
-export type TableItemNameProps = {
-  name: string;
-  description?: string;
-  descriptionIconVisible: boolean;
-};
-
-export function TableItemName({
-  name,
-  description,
-  descriptionIconVisible,
-}: TableItemNameProps) {
-  return (
-    <Flex alignItems="center" justifyContent="flex-start">
-      <Icon as={FiGrid} color="piperider.500" />
-      <Text mx={1}>{name}</Text>
-
-      {!descriptionIconVisible && (
-        <Tooltip label={description} placement="right-end" shouldWrapChildren>
-          <Icon as={FiAlertCircle} ml={1} />
-        </Tooltip>
-      )}
-    </Flex>
-  );
-}
-
-export function TableItemDescription({ description }: { description: string }) {
-  return (
-    <Text color="gray.500" noOfLines={3} textAlign="left">
-      <Text as="span">Description</Text>{' '}
-      <Text as="span" ml={4} title={description}>
-        {description}
-      </Text>
-    </Text>
+        <GridItem>
+          <TableItemName
+            name={tableName || ''}
+            description={description}
+            descriptionIconVisible={isExpanded}
+          />
+        </GridItem>
+        <GridItem>
+          <Flex color="gray.500" bg={'yellow.100'}>
+            <Text mr={4}>Rows</Text>
+            {/* DIFF_1: OK */}
+            {singleOnly ? (
+              <Text>
+                {formatColumnValueWith(fallbackTable?.row_count, formatNumber)}
+              </Text>
+            ) : (
+              // Rows 10 -> 10 | Columns 10 -> 10
+              <CRTableListDeltaSummary
+                baseCount={baseTableDatum?.row_count}
+                targetCount={targetTableDatum?.row_count}
+              />
+            )}
+          </Flex>
+        </GridItem>
+        <GridItem bg={'orange.100'}>
+          <Flex gap={2}>
+            {/* DIFF_2 : SR+CR */}
+            {singleOnly ? (
+              <AssertionLabel total={baseTotal} failed={baseFailed} />
+            ) : (
+              <CRTableListAssertions
+                baseAssertionFailed={baseFailed}
+                baseAssertionTotal={baseTotal}
+                targetAssertionFailed={targetFailed}
+                targetAssertionTotal={targetTotal}
+              />
+            )}
+            <Icon
+              position={'absolute'}
+              right={0}
+              ml={5}
+              as={isExpanded ? FiChevronUp : FiChevronDown}
+              color="piperider.500"
+              boxSize={6}
+            />
+          </Flex>
+        </GridItem>
+        <GridItem>
+          <Link
+            data-cy="cr-navigate-report-detail"
+            onClick={(event) => {
+              event.stopPropagation();
+              onSelect({ tableName });
+            }}
+          >
+            <Text
+              p={1}
+              borderRadius={'lg'}
+              border={'1px solid'}
+              borderColor={'blue.400'}
+              color={'blue.400'}
+            >
+              Table Details
+            </Text>
+          </Link>
+        </GridItem>
+        <GridItem colSpan={2}>
+          {isExpanded ? (
+            <TableItemDescription description={description || ''} />
+          ) : (
+            <Flex color="gray.500" maxWidth="650px" bg={'cyan.100'}>
+              <Text as="span" mr={4}>
+                Columns
+              </Text>
+              {/* DIFF_3 */}
+              {singleOnly ? (
+                <Flex
+                  __css={{
+                    display: 'flex',
+                    gap: 3,
+                    alignItems: 'center',
+                    maxWidth: '100%',
+                    overflowX: 'scroll',
+                    scrollbarWidth: 'none',
+                    '&::-webkit-scrollbar': {
+                      display: 'none',
+                    },
+                  }}
+                >
+                  {columns.length > 0 &&
+                    columns.map((name) => (
+                      <SRTableListColumnLabel
+                        key={name}
+                        name={name}
+                        icon={
+                          getIconForColumnType(baseTableDatum?.columns[name])
+                            .icon
+                        }
+                      />
+                    ))}
+                </Flex>
+              ) : (
+                <CRTableListColumnsSummary
+                  baseCount={baseTableDatum?.col_count}
+                  targetCount={targetTableDatum?.col_count}
+                />
+              )}
+            </Flex>
+          )}
+        </GridItem>
+      </Grid>
+    </TableAccordionWrapper>
   );
 }
