+++ static_report/src/utils/layout.ts
@@ -3,3 +3,4 @@ export const breadcrumbHeight = 36;
 export const mainContentAreaHeight = `calc(100vh - ${
   topNavAndFooterHeightOffset + breadcrumbHeight
 }px)`;
+export const tableListGridTempCols = '1fr 2fr 1.5fr';
